corre bem com zero users

+++ src/main/java/dei/uc/pt/aor/RunWSClient.java
@@ -33,8 +33,7 @@ public class RunWSClient {
 			try {
 				answer = reader.readLine();
 			} catch (IOException e) {
-				// usar log
-				e.printStackTrace();
+				System.out.println("Error reading input.");
 				stop = true;
 				answer = "0";
 			}
@@ -49,27 +48,27 @@ public class RunWSClient {
 			}
 
 			switch (op) {
-			case 1: 
+			case 1: // Manage users
 				menuUser();
 				break;
-			case 2: 
+			case 2: // Manage playlists
 				menuPlaylist();
 				break;
-			case 3: 
+			case 3: // Manage songs
 				menuSong();
 				break;
-			case 0:
+			case 0: // Exit application
 				System.out.println("Exiting...");
 				try {
 					reader.close();
 				} catch (IOException e) {
-					// usar log
-					e.printStackTrace();
+					System.out.println("Error closing BufferedReader.");
 				}
 				System.out.println("Goodbye!");
 				stop = true; 
 				break;
-			default: stop = false;
+			default: // Repeat menu
+				stop = false;
 			}
 		}
 
@@ -101,8 +100,7 @@ public class RunWSClient {
 			try {
 				answer = reader.readLine();
 			} catch (IOException e) {
-				// usar log
-				e.printStackTrace();
+				System.out.println("Error reading input.");
 				answer = "100";
 			}
 
@@ -122,34 +120,42 @@ public class RunWSClient {
 
 			try {
 				switch (op) {
-				case 1:
+				case 1: // Total number of users
 					int total = WSConsumerUser.totalUsers();
 					if (total >= 0) {
 						System.out.print("Total number of users: ");
 						System.out.println(total);
 					}
 					break;
-				case 2: WSConsumerUser.listAllUsers(); break;
-				case 3: 
-					System.out.print("Insert the user id or email: ");
-					while (!done) {
-						data1 = reader.readLine();
-						if (data1.matches("\\d+")) {
-							done = true;
-							WSpath = "/id";
-						} else if (data1.matches(".+@.+\\.[a-z]+")) {
-							done = true;
-							WSpath = "/email";
-						}
-						else System.out.print("Please insert a valid id or email: ");
-					}
-					WSConsumerUser.userInfo(data1, WSpath);
+				case 2: // List all users
+					WSConsumerUser.listAllUsers();
 					break;
-				case 4: //WSConsumerUser.totalLoggedUsers(); 
+				case 3: // Consult user info
+					total = WSConsumerUser.totalUsers();
+					if (total > 0) {
+						//list all users, choose one
+						System.out.print("Insert the user id or email: ");
+						while (!done) {
+							data1 = reader.readLine();
+							if (data1.matches("\\d+")) {
+								done = true;
+								WSpath = "/id";
+							} else if (data1.matches(".+@.+\\.[a-z]+")) {
+								done = true;
+								WSpath = "/email";
+							}
+							else System.out.print("Please insert a valid id or email: ");
+						}
+						WSConsumerUser.userInfo(data1, WSpath);
+					} else if (total == 0) System.out.println("***** There are no users *****");
+					break;
+				case 4: // Total number of logged users //check first
+					//WSConsumerUser.totalLoggedUsers(); 
 					System.out.println("Not implemented yet"); break;
-				case 5: //WSConsumerUser.listLoggedUsers(); break;
+				case 5: // List all logged users
+					//WSConsumerUser.listLoggedUsers(); break;
 					System.out.println("Not implemented yet"); break;
-				case 6: 
+				case 6: // Create new user
 					System.out.print("Insert the name: ");
 					data1 = reader.readLine();
 					while (!done) {
@@ -161,47 +167,56 @@ public class RunWSClient {
 						} else System.out.println("Please insert a valid email: ");
 					}
 					break;
-				case 7: 
-					System.out.print("Insert the user id or email: ");
-					while (!done) {
-						data1 = reader.readLine();
-						if (data1.matches("\\d+")) {
-							done = true;
-							WSpath = "/id";
-						} else if (data1.matches(".+@.+\\.[a-z]+")) {
-							done = true;
-							WSpath = "/email";
+				case 7: // Delete user
+					total = WSConsumerUser.totalUsers();
+					if (total > 0) {
+						//list all users, choose one
+						System.out.print("Insert the user id or email: ");
+						while (!done) {
+							data1 = reader.readLine();
+							if (data1.matches("\\d+")) {
+								done = true;
+								WSpath = "/id";
+							} else if (data1.matches(".+@.+\\.[a-z]+")) {
+								done = true;
+								WSpath = "/email";
+							}
+							else System.out.print("Please insert a valid id or email: ");
 						}
-						else System.out.print("Please insert a valid id or email: ");
-					}
-					System.out.println("Are you sure you want to delete user with id "+data1+" (y/n)?");
-					data2 = reader.readLine();
-					if ( (data2.charAt(0)=='Y') || (data2.charAt(0)=='y') )
-						WSConsumerUser.deleteUser(data1, WSpath);
+						System.out.println("Are you sure you want to delete user with id "+data1+" (y/n)?");
+						data2 = reader.readLine();
+						if ( (data2.charAt(0)=='Y') || (data2.charAt(0)=='y') )
+							WSConsumerUser.deleteUser(data1, WSpath);
+					} else if (total == 0) System.out.println("***** There are no users *****");
 					break;
-				case 8: 
-					System.out.print("Insert the user id or email: ");
-					while (!done) {
-						data1 = reader.readLine();
-						if (data1.matches("\\d+")) {
-							done = true;
-							WSpath = "/id";
-						} else if (data1.matches(".+@.+\\.[a-z]+")) {
-							done = true;
-							WSpath = "/email";
+				case 8: // Change user pass
+					total = WSConsumerUser.totalUsers();
+					if (total > 0) {
+						System.out.print("Insert the user id or email: ");
+						while (!done) {
+							data1 = reader.readLine();
+							if (data1.matches("\\d+")) {
+								done = true;
+								WSpath = "/id";
+							} else if (data1.matches(".+@.+\\.[a-z]+")) {
+								done = true;
+								WSpath = "/email";
+							}
+							else System.out.print("Please insert a valid id or email: ");
 						}
-						else System.out.print("Please insert a valid id or email: ");
-					}
-					System.out.println("Insert new password: ");
-					data2 = reader.readLine();
-					WSConsumerUser.changeUserPass(data1, data2, WSpath);
+						System.out.println("Insert new password: ");
+						data2 = reader.readLine();
+						WSConsumerUser.changeUserPass(data1, data2, WSpath);
+					} else if (total == 0) System.out.println("***** There are no users *****");
 					break;
-				case 0: stop = true; break;
-				default: stop = false;
+				case 0: // Go back
+					stop = true; 
+					break;
+				default: // Repeat menu
+					stop = false;
 				}
 			} catch (Exception e) {
-				// usar log
-				e.printStackTrace();
+				System.out.println("Error reading input.");
 				stop = true;
 			}
 
@@ -230,8 +245,7 @@ public class RunWSClient {
 			try {
 				answer = reader.readLine();
 			} catch (IOException e) {
-				// usar log
-				e.printStackTrace();
+				System.out.println("Error reading input.");
 				answer = "100";
 			}
 
@@ -251,15 +265,17 @@ public class RunWSClient {
 
 			try {
 				switch (op) {
-				case 1:
+				case 1: // Total number of playlists
 					int total =  WSConsumerPlaylist.totalPlaylists();
 					if (total >= 0) {
 						System.out.print("Total number of playlists: ");
 						System.out.println(total);
 					}
 					break;
-				case 2: WSConsumerPlaylist.listAllPlaylists(); break;
-				case 3: 
+				case 2: // List all playlists
+					WSConsumerPlaylist.listAllPlaylists();
+					break;
+				case 3: // List playlists of user
 					total = WSConsumerUser.totalUsers();
 					if (total > 0) {
 						if (WSConsumerUser.listAllUsers()) {
@@ -279,9 +295,9 @@ public class RunWSClient {
 							if (WSConsumerUser.userInfo(data, WSpath))
 								WSConsumerPlaylist.playlistsOfUser(data, WSpath);
 						}
-					} else if (total == 0) System.out.println("There are no users");
+					} else if (total == 0) System.out.println("***** There are no users *****");
 					break;
-				case 4: 
+				case 4: // Add/Remove songs to playlist
 					total = WSConsumerPlaylist.totalPlaylists();
 					if (total > 0) {
 						//list all playlists, choose one
@@ -298,14 +314,16 @@ public class RunWSClient {
 							if (WSConsumerPlaylist.playlistInfo(pid))
 								menuPlaylist(pid);
 						}
-					} else if (total == 0) System.out.println("There are no playlists");
+					} else if (total == 0) System.out.println("***** There are no playlists *****");
 					break;
-				case 0: stop = true; break;
-				default: stop = false;
+				case 0: // Go back
+					stop = true;
+					break;
+				default: // Repeat menu
+					stop = false;
 				}
 			} catch (Exception e) {
-				// usar log
-				e.printStackTrace();
+				System.out.println("Error reading input.");
 				stop = true;
 			}
 
@@ -333,8 +351,7 @@ public class RunWSClient {
 			try {
 				answer = reader.readLine();
 			} catch (IOException e) {
-				// usar log
-				e.printStackTrace();
+				System.out.println("Error reading input.");
 				answer = "100";
 			}
 
@@ -353,8 +370,10 @@ public class RunWSClient {
 
 			try {
 				switch (op) {
-				case 1: WSConsumerSong.songsOfPlaylist(pid); break;
-				case 2:
+				case 1: // List songs of playlist
+					WSConsumerSong.songsOfPlaylist(pid);
+					break;
+				case 2: // Add song to playlist
 					int total = WSConsumerSong.totalSongs();
 					if (total > 0) {
 						//list all songs, choose one to add
@@ -373,9 +392,9 @@ public class RunWSClient {
 								WSConsumerSong.songsOfPlaylist(pid);
 							}
 						}
-					} else if (total == 0) System.out.println("There are no songs");
+					} else if (total == 0) System.out.println("***** There are no songs *****");
 					break;
-				case 3: 
+				case 3: // Remove song from playlist
 					total = WSConsumerSong.totalSongsOfPlaylist(pid);
 					if (total > 0) {
 						//list all songs of playlist, choose one to remove
@@ -392,14 +411,15 @@ public class RunWSClient {
 							WSConsumerPlaylist.updateSongPlaylist(sid, pid, "/removesongfromplaylist");				
 							WSConsumerSong.songsOfPlaylist(pid);
 						}
-					} else if (total == 0) System.out.println("There are no songs in the playlist");
+					} else if (total == 0) System.out.println("***** There are no songs in the playlist *****");
 					break;
-				case 0: stop = true; break;
-				default: stop = false;
+				case 0: // Go back
+					stop = true; break;
+				default: // Repeat menu
+					stop = false;
 				}
 			} catch (Exception e) {
-				// usar log
-				e.printStackTrace();
+				System.out.println("Error reading input.");
 				stop = true;
 			}
 
@@ -428,8 +448,7 @@ public class RunWSClient {
 			try {
 				answer = reader.readLine();
 			} catch (IOException e) {
-				// usar log
-				e.printStackTrace();
+				System.out.println("Error reading input.");
 				answer = "100";
 			}
 
@@ -449,27 +468,31 @@ public class RunWSClient {
 
 			try {
 				switch (op) {
-				case 1: 
+				case 1: // Total number of songs
 					int total = WSConsumerSong.totalSongs();
 					if (total >= 0) {
 						System.out.print("Total number of songs: ");
 						System.out.println(total);
 					}
 					break;
-				case 2: WSConsumerSong.listAllSongs(); break;
-				case 3:
-					System.out.print("Insert the song id: ");
-					while (!done) {
-						data = reader.readLine();
-						if (data.matches("\\d+")) {
-							done = true;
-							id = Long.parseLong(data);
+				case 2: // List all songs
+					WSConsumerSong.listAllSongs(); break;
+				case 3: // Consult song info
+					total = WSConsumerSong.totalSongs();
+					if (total > 0) {
+						System.out.print("Insert the song id: ");
+						while (!done) {
+							data = reader.readLine();
+							if (data.matches("\\d+")) {
+								done = true;
+								id = Long.parseLong(data);
+							}
+							else System.out.print("Please insert a valid id: ");
 						}
-						else System.out.print("Please insert a valid id: ");
-					}
-					WSConsumerSong.songInfo(id);
+						WSConsumerSong.songInfo(id);
+					} else if (total == 0) System.out.println("***** There are no songs *****");
 					break;
-				case 4:
+				case 4: // Delete songs of user
 					total = WSConsumerUser.totalUsers();
 					if (total > 0) {
 						//list all users and choose one
@@ -491,17 +514,19 @@ public class RunWSClient {
 								int totals = WSConsumerSong.totalSongsOfUser(data, WSpath);
 								if (totals > 0)
 									menuUser(data, WSpath);
-								else if (totals == 0) System.out.println("The user has no songs.");
+								else if (totals == 0) System.out.println("***** The user has no songs *****");
 							}
 						}
-					} else if (total == 0) System.out.println("There are no users");
+					} else if (total == 0) System.out.println("***** There are no users *****");
 					break;
-				case 0: stop = true; break;
-				default: stop = false;
+				case 0: // Go back
+					stop = true;
+					break;
+				default: // Repeat menu
+					stop = false;
 				}
 			} catch (Exception e) {
-				// usar log
-				e.printStackTrace();
+				System.out.println("Error reading input.");
 				stop = true;
 			}
 
@@ -528,8 +553,7 @@ public class RunWSClient {
 			try {
 				answer = reader.readLine();
 			} catch (IOException e) {
-				// usar log
-				e.printStackTrace();
+				System.out.println("Error reading input.");
 				answer = "100";
 			}
 
@@ -548,8 +572,9 @@ public class RunWSClient {
 
 			try {
 				switch (op) {
-				case 1: WSConsumerSong.songsOfUser(idemail, WSpath); break;
-				case 2:
+				case 1: // List songs of user
+					WSConsumerSong.songsOfUser(idemail, WSpath); break;
+				case 2: // Delete song of user
 					//list all songs of user, choose one to remove
 					if (WSConsumerSong.songsOfUser(idemail, WSpath)) {
 						//choose song to delete
@@ -566,12 +591,14 @@ public class RunWSClient {
 							WSConsumerSong.songsOfUser(idemail, WSpath);
 					}
 					break;
-				case 0: stop = true; break;
-				default: stop = false;
+				case 0: // Go back
+					stop = true;
+					break;
+				default: // Repeat menu
+					stop = false;
 				}
 			} catch (Exception e) {
-				// usar log
-				e.printStackTrace();
+				System.out.println("Error reading input.");
 				stop = true;
 			}
